moved pi test ficture to a module
